requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.0.1
numpy==1.24.3
tensorflow==2.19.0
firebase-admin==6.2.0
